+++ app/controllers/api/decks_controller.rb
@@ -4,13 +4,17 @@ module Api
     before_action :set_options, only: %i[index show update create]
 
     def index
-      @decks = Deck.all
+      @decks = user_signed_in? ? Deck.where(user: current_user).or(Deck.where(public: true)) : Deck.where(public: true)
       render json:
                DeckSerializer.new(@decks, @options).serializable_hash.to_json
     end
 
     def show
-      render json: DeckSerializer.new(@deck, @options).serializable_hash.to_json
+      if @deck.public || (user_signed_in?) && @deck.user == current_user
+        render json: DeckSerializer.new(@deck, @options).serializable_hash.to_json
+      else
+        render json: {}, status: 401
+      end
     end
 
     def create
@@ -28,7 +32,7 @@ module Api
     end
 
     def update
-      if user_signed_in?
+      if user_signed_in? && @deck.user == current_user
         if @deck.update(deck_params)
           render json:
                    DeckSerializer.new(@deck, @options).serializable_hash.to_json
@@ -41,7 +45,7 @@ module Api
     end
 
     def destroy
-      if user_signed_in?
+      if user_signed_in? && @deck.user == current_user
         if @deck.destroy
           head :no_content
         else
